+++ packages/prime-core/__tests__/utils/createSchemaQuery.ts
@@ -1,4 +1,4 @@
-import { startCase } from 'lodash';
+import { camelCase, upperFirst } from 'lodash';
 
 const enumValue = value => {
   function enumValueOperator() {
@@ -41,7 +41,7 @@ const mapFields = fields => {
       },
       position
     ) => {
-      const title = passProps.title || startCase(name);
+      const title = passProps.title || upperFirst(camelCase(name));
       return {
         name,
         title,
@@ -63,7 +63,9 @@ export const createSchemaQuery = (name, fields) => {
     title: name,
     description: name,
     variant: enumValue('Default'),
-    settings: {},
+    settings: {
+      mutations: true,
+    },
     fields: [
       {
         title: 'Main',