+++ packages/prime-core/src/modules/internal/resolvers/DocumentResolver.ts
@@ -81,14 +81,14 @@ export class DocumentResolver {
 
   @Authorized()
   @Query(returns => DocumentConnection)
-  public allDocuments(
+  public async allDocuments(
     @Arg('sort', type => [DocumentSort], { defaultValue: 1, nullable: true }) sorts: string[],
     @Arg('filter', type => [DocumentFilterInput], { nullable: true })
     filters: DocumentFilterInput[],
     @Args() args: ConnectionArgs
   ) {
     const result = new ExtendedConnection(args, {
-      where: qb => {
+      where: (qb, counter = false) => {
         qb.andWhere('Document.deletedAt IS NULL');
 
         const subquery = qb
@@ -103,15 +103,23 @@ export class DocumentResolver {
             filterArr.map((filter, i) => {
               builder.orWhere(
                 new Brackets(sq => {
-                  Object.entries(filter).map(([key, value]) =>
-                    sq.andWhere(`${name}.${key} = :${key}_${i}`, { [`${key}_${i}`]: value })
-                  );
+                  Object.entries(filter).map(([key, value]) => {
+                    if (value === null) {
+                      sq.andWhere(`${name}.${key} IS NULL`);
+                    } else {
+                      sq.andWhere(`${name}.${key} = :${key}_${i}`, { [`${key}_${i}`]: value });
+                    }
+                  });
                 })
               );
             });
           });
 
-        const filtered = (filters || []).map(filter => pickBy(filter, identity));
+        const filtered = (filters || []).map(filter =>
+          pickBy(filter, (value, key) => {
+            return key === 'releaseId' || identity(value);
+          })
+        );
 
         if (filtered.length > 0 && Object.keys(filtered[0]).length > 0) {
           subquery.andWhere(filterWithName('d', filtered));
@@ -120,15 +128,16 @@ export class DocumentResolver {
 
         subquery.orderBy({ 'd.createdAt': 'DESC' }).limit(1);
 
-        qb.having(`Document.id = ${subquery.getQuery()}`);
-        qb.groupBy('Document.id');
+        if (!counter) {
+          qb.having(`Document.id = ${subquery.getQuery()}`);
+          qb.groupBy('Document.id');
+        }
         return qb;
       },
       repository: this.documentRepository,
       sortOptions: sortOptions(sorts),
     });
-
-    (result as any).totalCount = 15;
+    result.totalCountField = 'documentId';
 
     return result;
   }
@@ -248,6 +257,7 @@ export class DocumentResolver {
   public async versions(@Root() document: Document): Promise<Document[]> {
     return this.documentRepository.find({
       where: { documentId: document.documentId, deletedAt: IsNull() },
+      order: { updatedAt: 'DESC' },
     });
   }