+++ packages/prime-core/src/utils/config.ts
@@ -15,6 +15,7 @@ export const config = rc('prime', {
     '@primecms/field-slice',
     '@primecms/field-string',
   ],
+  sofaApi: true,
   uiDir: defaultUiDir(),
   coreUrl: defaultCoreUrl(),
 });