+++ packages/prime-field-boolean/src/ui/SchemaSettingsComponent.tsx
@@ -1,8 +1,23 @@
 import { PrimeFieldProps } from '@primecms/field';
+import { Form, Input, Switch } from 'antd';
 import React from 'react';
 
-export class SchemaSettingsComponent extends React.PureComponent<PrimeFieldProps> {
+type Props = PrimeFieldProps & {
+  options?: { label: string; default: boolean };
+};
+
+export class SchemaSettingsComponent extends React.PureComponent<Props> {
   public render() {
-    return null;
+    const { form } = this.props;
+    return (
+      <>
+        <Form.Item label="On by default">
+          {form.getFieldDecorator('options.default', {
+            valuePropName: 'checked',
+          })(<Switch />)}
+        </Form.Item>
+        <Form.Item label="Label">{form.getFieldDecorator('options.label')(<Input />)}</Form.Item>
+      </>
+    );
   }
 }