+++ packages/prime-field-group/src/ui/InputComponent.tsx
@@ -3,9 +3,26 @@ import { Button, Card, Form, Icon } from 'antd';
 import { get } from 'lodash';
 import React from 'react';
 
-const { uuid } = window as any;
+const randomByte = () => {
+  const seq = Math.floor(Math.random() * Number.MAX_SAFE_INTEGER).toString(16);
+  return seq.substr(seq.length - 1, 1);
+};
+const randomBytes = length =>
+  Array.from({ length })
+    .map(randomByte)
+    .join('');
+const randomUuid = () => [8, 4, 4, 4, 12].map(randomBytes).join('-');
 
-const getItems = ({ initialValue }: PrimeFieldProps) => {
+const { uuid = { v4: randomUuid } } = window as any;
+
+const initialToken = uuid.v4();
+
+const getItems = ({ initialValue, field }: PrimeFieldProps) => {
+  const options = { ...field.defaultOptions, ...field.options };
+  if (!options.repeated) {
+    return [[initialToken, 0]];
+  }
+
   if (Array.isArray(initialValue)) {
     return initialValue.map((_, index) => [uuid.v4(), index]);
   }
@@ -22,6 +39,11 @@ export class InputComponent extends React.PureComponent<PrimeFieldProps, any> {
     items: getItems(this.props),
     index: getIndex(this.props),
   };
+
+  get options() {
+    const { field } = this.props;
+    return { ...field.defaultOptions, ...field.options };
+  }
 
   public componentWillReceiveProps(nextProps: PrimeFieldProps) {
     if (!this.props.document && nextProps.document) {
@@ -61,7 +83,7 @@ export class InputComponent extends React.PureComponent<PrimeFieldProps, any> {
   };
 
   public renderField = (field: any, key: string, index: number) => {
-    const repeated = get(this.props.field, 'options.repeated', false);
+    const repeated = this.options.repeated;
     const prefix = repeated ? `${index}.` : '';
     const path = `${this.props.path}.${prefix}${field.name}`;
     const initialValue = get(this.props.initialValue, `${prefix}${field.name}`);
@@ -77,7 +99,7 @@ export class InputComponent extends React.PureComponent<PrimeFieldProps, any> {
   public renderGroupItem = ([key, index]: any) => {
     const { field } = this.props;
     const { fields = [] } = field;
-    const repeated = get(field, 'options.repeated', false);
+    const repeated = this.options.repeated;
 
     if (!key) {
       return null;
@@ -103,7 +125,11 @@ export class InputComponent extends React.PureComponent<PrimeFieldProps, any> {
   public render() {
     const { items } = this.state;
     const { field } = this.props;
-    const repeated = get(field, 'options.repeated', false);
+    const repeated = this.options.repeated;
+
+    if (field.fields.length === 0) {
+      return null;
+    }
 
     return (
       <Form.Item label={field.title} className="prime-group">