+++ packages/prime-ui/src/routes/documents/DocumentsDetail.tsx
@@ -226,16 +226,23 @@ export class DocumentsDetail extends React.Component<IProps> {
 
   public onKeyDown = (e: any) => {
     if (e.which === 83 && (e.ctrlKey || e.metaKey)) {
-      return this.onSave(e);
+      if (e.shiftKey) {
+        return this.onPublish(e);
+      } else {
+        return this.onSave(e);
+      }
     }
     return true;
   };
 
-  public onSave = async (e: React.MouseEvent<HTMLElement>) => {
+  public onSave = async (e: React.MouseEvent<HTMLElement>, feedback = true) => {
     e.preventDefault();
     try {
-      await this.save({ documentId: this.props.match.params.entryId });
-      message.success('Document was saved', 1);
+      const releaseId = this.options.release;
+      await this.save({ documentId: this.props.match.params.entryId, releaseId });
+      if (feedback) {
+        message.success('Document was saved', 1);
+      }
     } catch (err) {
       message.error('Could not save document');
       console.error(err); // tslint:disable-line no-console
@@ -246,6 +253,9 @@ export class DocumentsDetail extends React.Component<IProps> {
   public onPublish = async (e: any) => {
     if (this.contentEntry) {
       this.loading.publish = true;
+      if (this.promptEnabled && this.documentForm!.props.form.isFieldsTouched()) {
+        await this.onSave(e, false);
+      }
       await this.contentEntry.publish();
       message.success('Document was published', 1);
       this.loading.publish = false;