+++ packages/prime-field-group/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3-beta.4](https://github.com/birkir/prime/tree/master/packages/prime-field-group/compare/v0.3.3-beta.3...v0.3.3-beta.4) (2019-03-01)
+
+### Bug Fixes
+
+- group fields and related issues ([#103](https://github.com/birkir/prime/tree/master/packages/prime-field-group/issues/103)) ([ca95718](https://github.com/birkir/prime/tree/master/packages/prime-field-group/commit/ca95718))
+
+## [0.3.3-beta.3](https://github.com/birkir/prime/tree/master/packages/prime-field-group/compare/v0.3.3-beta.2...v0.3.3-beta.3) (2019-02-20)
+
+### Bug Fixes
+
+- fields general ([#93](https://github.com/birkir/prime/tree/master/packages/prime-field-group/issues/93)) ([94edde6](https://github.com/birkir/prime/tree/master/packages/prime-field-group/commit/94edde6))
+
+## [0.3.3-beta.2](https://github.com/birkir/prime/tree/master/packages/prime-field-group/compare/v0.3.3-beta.1...v0.3.3-beta.2) (2019-02-18)
+
+### Bug Fixes
+
+- graphql name startCase bug ([3b8e46f](https://github.com/birkir/prime/tree/master/packages/prime-field-group/commit/3b8e46f))
+
+## [0.3.3-beta.1](https://github.com/birkir/prime/tree/master/packages/prime-field-group/compare/v0.3.3-beta.0...v0.3.3-beta.1) (2019-02-15)
+
+**Note:** Version bump only for package @primecms/field-group
+
 ## [0.3.3-beta.0](https://github.com/birkir/prime/tree/master/packages/prime-field-group/compare/v0.3.2-beta.9...v0.3.3-beta.0) (2019-02-12)
 
 **Note:** Version bump only for package @primecms/field-group