+++ packages/prime-field-string/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3-beta.4](https://github.com/birkir/prime/tree/master/packages/prime-field-string/compare/v0.3.3-beta.3...v0.3.3-beta.4) (2019-03-01)
+
+**Note:** Version bump only for package @primecms/field-string
+
+## [0.3.3-beta.2](https://github.com/birkir/prime/tree/master/packages/prime-field-string/compare/v0.3.3-beta.1...v0.3.3-beta.2) (2019-02-18)
+
+**Note:** Version bump only for package @primecms/field-string
+
+## [0.3.3-beta.1](https://github.com/birkir/prime/tree/master/packages/prime-field-string/compare/v0.3.3-beta.0...v0.3.3-beta.1) (2019-02-15)
+
+**Note:** Version bump only for package @primecms/field-string
+
 ## [0.3.3-beta.0](https://github.com/birkir/prime/tree/master/packages/prime-field-string/compare/v0.3.2-beta.9...v0.3.3-beta.0) (2019-02-12)
 
 **Note:** Version bump only for package @primecms/field-string