+++ packages/prime-ui/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3-beta.4](https://github.com/birkir/prime/compare/v0.3.3-beta.3...v0.3.3-beta.4) (2019-03-01)
+
+### Bug Fixes
+
+- group fields and related issues ([#103](https://github.com/birkir/prime/issues/103)) ([ca95718](https://github.com/birkir/prime/commit/ca95718))
+
+## [0.3.3-beta.3](https://github.com/birkir/prime/compare/v0.3.3-beta.2...v0.3.3-beta.3) (2019-02-20)
+
+### Bug Fixes
+
+- fields general ([#93](https://github.com/birkir/prime/issues/93)) ([94edde6](https://github.com/birkir/prime/commit/94edde6))
+
+## [0.3.3-beta.2](https://github.com/birkir/prime/compare/v0.3.3-beta.1...v0.3.3-beta.2) (2019-02-18)
+
+**Note:** Version bump only for package @primecms/ui
+
+## [0.3.3-beta.1](https://github.com/birkir/prime/compare/v0.3.3-beta.0...v0.3.3-beta.1) (2019-02-15)
+
+### Bug Fixes
+
+- onboarding ([08dd305](https://github.com/birkir/prime/commit/08dd305))
+- publish release doc count ([b84181c](https://github.com/birkir/prime/commit/b84181c))
+- releases and connection counts ([bf86d25](https://github.com/birkir/prime/commit/bf86d25))
+- webhook calls ([1740dac](https://github.com/birkir/prime/commit/1740dac))
+
 ## [0.3.3-beta.0](https://github.com/birkir/prime/compare/v0.3.2-beta.9...v0.3.3-beta.0) (2019-02-12)
 
 **Note:** Version bump only for package @primecms/ui